# docs/swagger.yaml
basePath: /api/pve
definitions:
  main.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  main.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/main.ErrorDetail'
    type: object
  main.OperationErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  main.OperationResponse:
    properties:
      error:
        $ref: '#/definitions/main.OperationErrorBody'
      id:
        type: string
      resourceLocation:
        type: string
      status:
        type: string
    type: object
  pve.CreateVmRequest:
    properties:
      aptMirror:
        description: APT 镜像源地址，写入 cloud-init apt.primary（可选，仅 ciPackages 非空时生效）
        example: http://mirrors.aliyun.com/ubuntu
        type: string
      ciPackages:
        description: CloudInit 首次开机需安装的软件包列表（可选；非空时须同时指定 snippetsStorage）
        example:
        - qemu-guest-agent
        items:
          type: string
        type: array
      ciPassword:
        description: CloudInit 登录密码（可选）
        example: secret
        type: string
      ciUser:
        description: CloudInit 配置（可选，使用云镜像时配置首次启动参数）
        example: ubuntu
        type: string
      cores:
        description: CPU 核数（必填）
        example: 2
        type: integer
      diskFormat:
        description: 源磁盘格式，如 qcow2/raw（可选）
        example: qcow2
        type: string
      diskInterface:
        description: 磁盘接口类型，默认 virtio0（可选）
        example: scsi0
        type: string
      diskSource:
        description: 导入磁盘来源，格式 storage:path（必填）
        example: local:import/noble-server-cloudimg-amd64.qcow2
        type: string
      ipConfig0:
        description: CloudInit 网络配置，格式同 PVE ipconfig0，如 ip=dhcp（可选）
        example: ip=192.168.1.100/24,gw=192.168.1.1
        type: string
      memory:
        description: 内存大小，单位 MB（必填）
        example: 2048
        type: integer
      name:
        description: 虚拟机名称（必填）
        example: my-vm
        type: string
      nameserver:
        description: CloudInit DNS 服务器地址（可选）
        example: 8.8.8.8
        type: string
      network:
        description: 网络设备配置，格式同 PVE net0，默认 virtio,bridge=vmbr0（可选）
        example: virtio,bridge=vmbr0
        type: string
      searchDomain:
        description: CloudInit DNS 搜索域（可选）
        example: example.com
        type: string
      snippetsStorage:
        description: 存放 cloud-init user-data Snippet 的 PVE 存储名称（ciPackages 非空时必填）
        example: local
        type: string
      sshKeys:
        description: CloudInit SSH 公钥，多个公钥用换行分隔（可选）
        example: ssh-rsa AAAA...
        type: string
      storage:
        description: 目标存储池（必填）
        example: local-lvm
        type: string
      vmid:
        description: 新虚拟机 VMID（必填）
        example: 200
        type: integer
    type: object
  pve.VM:
    properties:
      cpus:
        type: integer
      maxmem:
        type: integer
      mem:
        type: integer
      name:
        type: string
      status:
        type: string
      vmid:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Hyperflow PVE 管理接口
  title: Hyperflow API
  version: "1.0"
paths:
  /nodes:
    get:
      description: 返回 PVE 集群中所有节点列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 列出所有节点
      tags:
      - nodes
  /nodes/{node}:
    get:
      description: 返回指定节点的详细状态信息
      parameters:
      - description: 节点名称
        in: path
        name: node
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 获取节点详情
      tags:
      - nodes
  /nodes/{node}/vms:
    get:
      description: 返回指定节点上所有 QEMU 虚拟机列表
      parameters:
      - description: 节点名称
        in: path
        name: node
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pve.VM'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 列出节点上所有虚拟机
      tags:
      - vms
    post:
      consumes:
      - application/json
      description: |-
        通过 PVE 创建新虚拟机，并在创建时通过 import-from 导入指定磁盘卷。
        支持可选的 CloudInit 配置（ciUser、ciPassword、sshKeys、ipConfig0、nameserver、searchDomain）；
        当请求体包含任意 CloudInit 字段时，系统自动附加 CloudInit 驱动盘（ide2）及对应配置。
        ciPackages 非空时在 snippetsStorage 中生成 cloud-init user-data Snippet（自动执行 package_update/upgrade，并将 ciUser、ciPassword、sshKeys 写入 user-data）并通过 cicustom 引用（snippetsStorage 此时必填）。
        成功后响应 Operation-Location 及 Location 头。
      parameters:
      - description: 节点名称
        in: path
        name: node
        required: true
        type: string
      - description: 创建参数（vmid、cores、memory、diskSource、storage 为必填；name 为可选，不填时自动生成随机主机名；CloudInit
          字段均为可选；ciPackages 非空时 snippetsStorage 必填）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pve.CreateVmRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 新建虚拟机并导入磁盘
      tags:
      - vms
  /nodes/{node}/vms/{vmid}:
    delete:
      description: 异步删除指定虚拟机（虚拟机须处于停止状态），返回 LRO Operation-Location header
      parameters:
      - description: 节点名称
        in: path
        name: node
        required: true
        type: string
      - description: 虚拟机 ID
        in: path
        name: vmid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 删除虚拟机
      tags:
      - vms
    get:
      description: 返回指定虚拟机的当前状态信息
      parameters:
      - description: 节点名称
        in: path
        name: node
        required: true
        type: string
      - description: 虚拟机 ID
        in: path
        name: vmid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 获取指定虚拟机信息
      tags:
      - vms
  /nodes/{node}/vms/{vmid}/start:
    post:
      description: 异步启动指定虚拟机，返回 LRO Operation-Location header
      parameters:
      - description: 节点名称
        in: path
        name: node
        required: true
        type: string
      - description: 虚拟机 ID
        in: path
        name: vmid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 启动虚拟机
      tags:
      - vms
  /nodes/{node}/vms/{vmid}/stop:
    post:
      description: 异步停止指定虚拟机，返回 LRO Operation-Location header
      parameters:
      - description: 节点名称
        in: path
        name: node
        required: true
        type: string
      - description: 虚拟机 ID
        in: path
        name: vmid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 停止虚拟机
      tags:
      - vms
  /operations/{id}:
    get:
      description: 返回指定异步操作的当前状态，遵循微软 LRO 规范。status 为 Running / Succeeded / Failed
        之一。
      parameters:
      - description: 操作 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.OperationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 查询异步操作状态
      tags:
      - operations
  /storage:
    get:
      description: 返回 PVE 中所有存储资源列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: 列出所有存储
      tags:
      - storage
swagger: "2.0"